+++ src/board/move_maker.h
@@ -18,6 +18,9 @@ class CMoveMaker {
     void unmake_all();
     bool play_variation(const std::string &variation);
     void reset_history();
+  public:
+    void make_null_move();
+    void unmake_null_move();
   private:
     void update_castling_rights(const std::string &textual_move) const;
     void update_castling_rights(const SSquare source_or_target_square) const;