+++ src/move_generator/move_list.cpp
@@ -11,6 +11,21 @@
 
 constexpr unsigned int NOT_FOUND = INT_MAX;
 
+bool square_attacked_by_side_to_move(const SSquare square) {
+       CMoveGenerator response_generator;
+       response_generator.generate_all();
+        response_generator.move_list.prune_silent_moves();
+        response_generator.move_list.filter_captures_by_target_square(square);
+    return (response_generator.move_list.list_size() > 0);
+}
+
+bool square_attacked_by_side_not_to_move(const SSquare square) {
+    board.move_maker.make_null_move();
+    bool result = square_attacked_by_side_to_move(square);
+    board.move_maker.unmake_null_move();
+    return result;
+}
+
 CMoveList::CMoveList() {
     clear();
 }
@@ -70,7 +85,6 @@ SMove CMoveList::lookup_move(const std::string &text_move) const {
     if (index == NOT_FOUND) {
         return NULL_MOVE;
     }
-    assert(index >= 0);
     assert(index <  LIST_SIZE);
     return bidirectional_move_list[index];
 }
@@ -173,6 +187,7 @@ void CMoveList::store_silent_move(const SMove &move) {
 }
 
 void CMoveList::store_capture(const SMove &move) {
+    // If this assertion fails, then we need to adjust MAX_CAPTURES_IN_CHESS_POSITION
     assert(first_capture > 0);
     assert(consumer_position >= first_capture);
     --first_capture;
@@ -192,15 +207,11 @@ void CMoveList::prune_illegal_moves() {
     while (move != NULL_MOVE) {
         board.move_maker.make_move(move);
         SSquare my_king_square = CBoardLogic::king_square(my_colour);
-       CMoveGenerator response_generator;
-       response_generator.generate_all();
-        board.move_maker.unmake_move();
-        response_generator.move_list.prune_silent_moves();
-        response_generator.move_list.filter_captures_by_target_square(my_king_square);
-        if (response_generator.move_list.list_size() > 0) {
+        if (square_attacked_by_side_to_move(my_king_square)) {
             SMove illegal_move = move;
             remove(illegal_move);
         }
+        board.move_maker.unmake_move();
         move = get_next();
     }
     prune_illegal_castlings();
@@ -289,7 +300,9 @@ std::string CMoveList::as_text() const {
 
 void CMoveList::remove(const SMove move) {
     unsigned int position = get_index(move);
-    assert(position != NOT_FOUND);
+    if (position == NOT_FOUND) {
+        return;
+    }
     assert(position >= first_capture);
     assert(position < last_silent_move);
     if (position < LIST_ORIGIN) {
@@ -328,6 +341,10 @@ void CMoveList::prune_illegal_castlings() {
         if (!move_on_list("e1d1") && move_on_list("e1c1")) {
             remove("e1c1");
         }
+        if (square_attacked_by_side_not_to_move(king_square)) {
+            remove("e1g1");
+            remove("e1c1");
+        }
         return;
     }
     if (!move_on_list("e8f8") && move_on_list("e8g8")) {
@@ -336,5 +353,9 @@ void CMoveList::prune_illegal_castlings() {
     if (!move_on_list("e8d8") && move_on_list("e8c8")) {
         remove("e8c8");
     }
+        if (square_attacked_by_side_not_to_move(king_square)) {
+            remove("e8g8");
+            remove("e8c8");
+        }
 }