+++ src/move_generator/move_list.h
@@ -8,11 +8,11 @@
 #include "move.h"
 #include "../technical_functions/standard_headers.h"
 
-// https://www.chessprogramming.org/Chess_Position
+// https://www.talkchess.com/forum/viewtopic.php?t=39332
 constexpr int MAX_MOVES_IN_CHESS_POSITION = 218;
-// Just a guess!
-// TEMP!!! Problems with stackframe mismatch!
-constexpr unsigned int MAX_CAPTURES_IN_CHESS_POSITION = 25;
+// Captures: probably 74. We use conservative 80.
+// https://chess.stackexchange.com/questions/8323/maximum-number-of-captures-in-a-position
+constexpr unsigned int MAX_CAPTURES_IN_CHESS_POSITION = 80;
 // Our move-list is bidirectional; left part for captures, right part for silent moves.
 constexpr unsigned int LIST_SIZE = MAX_MOVES_IN_CHESS_POSITION + MAX_CAPTURES_IN_CHESS_POSITION;
 constexpr unsigned int LIST_ORIGIN = MAX_CAPTURES_IN_CHESS_POSITION;