+++ src/technical_functions/engine_test.cpp
@@ -14,6 +14,7 @@
 #include "../evaluator/test_piece_square_value_tables.h"
 #include "../move_generator/test_move_list.h"
 #include "../move_generator/test_move_generator.h"
+#include "../move_generator/test_perft.h"
 #include "../opening_book/test_opening_book.h"
 #include "../search/test_search.h"
 #include "../search/test_statistics.h"
@@ -26,6 +27,7 @@
     EXPECT(CTestBoardLogic::test_everything());
     EXPECT(CTestMoveList::test_everything());
     EXPECT(CTestMoveGenerator::test_everything());
+    EXPECT(CTestPerft::test_everything());
     EXPECT(CTestMoveMaker::test_everything());
     EXPECT(CTestPieceSquareValueTables::test_everything());
     EXPECT(CTestEvaluator::test_everything());